feat: the player can shoot once per second

Closes #21

See merge request epitech-it-2027/rts!23

+++ gameEngine/ECS/Managers/Component/StructComponent.hpp
@@ -11,8 +11,8 @@
 
 struct Position
 {
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 struct BoundingBox
@@ -23,8 +23,8 @@ struct BoundingBox
 
 struct Transform
 {
-    int velX;
-    int velY;
+    float velX;
+    float velY;
 };
 
 struct Player
